requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow==2.13.0
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.76
scikit-learn==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
networkx==3.1
python-multipart==0.0.6
notebook>=6.4.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
Pillow>=9.0.0
opencv-python>=4.5.0
pandas>=1.3.0
psutil>=5.9.0